requirements.txt
appdirs==1.4.4
audioread==2.1.9
beautifulsoup4==4.9.3
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
decorator==4.4.2
dialogflow==1.1.0
docopt==0.6.2
gcloud==0.17.0
google-api-core==1.24.1
google-auth==1.24.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
httplib2==0.18.1
idna==2.10
joblib==1.0.0
Js2Py==0.70
jws==0.1.3
keyboard==0.13.5
librosa==0.8.0
llvmlite==0.35.0
numpy==1.19.3
oauth2client==3.0.0
packaging==20.8
pipwin==0.5.0
pooch==1.3.0
protobuf==4.0.0rc2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycparser==2.20
pycryptodome==3.4.3
pygame==2.0.1
pyjsparser==2.7.1
pyparsing==2.4.7
PyPrind==2.11.2
Pyrebase==3.0.27
pySmartDL==1.3.4
python-jwt==2.0.1
pytz==2020.5

requests-toolbelt==0.7.0
resampy==0.2.2
rsa==4.6
scikit-learn==0.24.0
scipy==1.5.4
six==1.15.0
SoundFile==0.10.3.post1
soupsieve==2.1
SpeechRecognition==3.8.1
threadpoolctl==2.1.0
tzlocal==2.1
urllib3==1.26.2
